add features
1. add version number
2. specific exception for error handling
3. move word list to new file
