+++ setting.py
@@ -16,7 +16,7 @@ MAX_NONCE = 4294967295
 MAX_TX_HISTORY_LIMIT = 10000
 MIN_CONTEND_SLOT_BALANCE = 1000000 * VEE
 
-THROW_EXCEPTION_ON_ERROR = False
+THROW_EXCEPTION_ON_ERROR = True
 CHECK_FEE_SCALE = True
 
 DEFAULT_CHAIN = 'mainnet'
